Table view Values are Centered

+++ src/main/java/lk/ijse/PastryPal/dto/ProductDto.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 public class ProductDto {
     private String product_id;
     private String description;
-    private double qty;
+    private int qty;
     private double price;
 }
